+++ ingeniamotion/communication.py
@@ -14,6 +14,7 @@ import ingenialogger
 from ingenialink.canopen.network import CAN_BAUDRATE, CAN_CHANNELS, CAN_DEVICE, CanopenNetwork
 from ingenialink.canopen.servo import CanopenServo
 from ingenialink.dictionary import Interface
+from ingenialink.emcy import EmergencyMessage
 from ingenialink.enums.register import REG_ACCESS, REG_DTYPE
 from ingenialink.enums.servo import SERVO_STATE
 from ingenialink.eoe.network import EoENetwork
@@ -989,6 +990,40 @@ class Communication(metaclass=MCMetaClass):
         """
         drive = self.mc._get_drive(servo)
         drive.unsubscribe_from_status(callback)
+
+    def subscribe_emergency_message(
+        self,
+        callback: Callable[[EmergencyMessage], None],
+        servo: str = DEFAULT_SERVO,
+    ) -> None:
+        """Subscribe to emergency messages.
+
+        Only available for CANopen and EtherCAT CoE protocols.
+
+        Args:
+            callback :  Callable that takes an EmergencyMessage instance as argument.
+            servo : servo alias to reference it. ``default`` by default.
+
+        """
+        drive = self.mc._get_drive(servo)
+        drive.emcy_subscribe(callback)
+
+    def unsubscribe_emergency_message(
+        self,
+        callback: Callable[[EmergencyMessage], None],
+        servo: str = DEFAULT_SERVO,
+    ) -> None:
+        """Unsubscribe from emergency messages.
+
+        Only available for CANopen and EtherCAT CoE protocols.
+
+        Args:
+            callback : Subscribed callback.
+            servo : servo alias to reference it. ``default`` by default.
+
+        """
+        drive = self.mc._get_drive(servo)
+        drive.emcy_unsubscribe(callback)
 
     def load_firmware_canopen(
         self,